% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_data.R
\name{en_data}
\alias{en_data}
\title{Fetch data}
\usage{
en_data(id = NULL, result = NULL, format = "xml", ...)
}
\arguments{
\item{id}{(character) Ids of various sorts, combine in a vector for > 1}

\item{result}{(character) result option, one of: sequence_release, assembly, ...}

\item{format}{(character) data format, one of xml (default), fasta, text, fastq}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
object of class \code{xml_document}
}
\description{
Fetch data
}
\examples{
\dontrun{
en_data(id = 'Taxon:4235')
en_data(id = c('Taxon:4235', 'Taxon:6543'))
en_data(id = 'A00145', format = "fasta")
en_data(id = 'A00145', format = "fastq")
en_data(id = 'A00145', format = "text")
## not working right now
# taxa <- c('Taxon:Human', 'Taxon:Cat', 'Taxon:Mouse', 'Taxon:Zebrafish')
# en_data(id = taxa)
}
}
\references{
\url{http://www.ebi.ac.uk/ena/browse/data-retrieval-rest}
}

