% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_download.R
\name{en_download}
\alias{en_download}
\title{Fetch data}
\usage{
en_download(id = NULL, result = NULL, format = "xml", file = NULL, ...)
}
\arguments{
\item{id}{Ids of various sorts, combine in a vector for > 1}

\item{result}{result option, one of: sequence_release, assembly, ...}

\item{format}{data format, one of xml (default), fasta, text, fastq}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
data.frame
}
\description{
Fetch data
}
\examples{
\dontrun{
en_download(id = 'Taxon:4235')
en_download(id = c('Taxon:4235', 'Taxon:6543'))
en_download(id = c('Taxon:Human', 'Taxon:Cat', 'Taxon:Mouse', 'Taxon:Zebrafish'))
}
}

