% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch.R
\name{en_fetch}
\alias{en_fetch}
\title{Fetch data}
\usage{
en_fetch(id = NULL, result = NULL, ...)
}
\arguments{
\item{id}{Ids of various sorts, combine in a vector for > 1}

\item{result}{result option, one of: sequence_release, assembly, ...}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
data.frame
}
\description{
Fetch data
}
\examples{
\dontrun{
en_fetch(id = 'Taxon:4235')
en_fetch(id = c('Taxon:4235', 'Taxon:6543'))
en_fetch(id = c('Taxon:Human', 'Taxon:Cat', 'Taxon:Mouse', 'Taxon:Zebrafish'))
}
}

