% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/en_search.R
\name{en_search}
\alias{en_search}
\title{Search for data}
\usage{
en_search(query = NULL, result = NULL, limit = NULL, offset = NULL,
  max = NULL, ...)
}
\arguments{
\item{query}{(character) Query terms}

\item{result}{(character) result option, one of: sequence_release,
assembly, coding_release, coding_update, read_run (are there others?)}

\item{limit}{(integer) Number of results to return}

\item{offset}{(integer) Result number to start at}

\item{max}{(integer) Max results}

\item{...}{Curl options passed to \code{\link[httr]{GET}}}
}
\value{
A list of lists, where each nested list is a record. in the future
we hope to parse to a single data.frame
}
\description{
Search for data
}
\examples{
\dontrun{
# sequence_release
res <- en_search(query = '"geo_circ(-0.587,-90.5713,170)"', result='sequence_release')
lapply(res, "[[", "seq")
## another eg.
en_search(query = '"geo_circ(-0.587,-90.5713,170)"', result='sequence_release', limit = 1)

# assembly
res <- en_search(query = '"tax_eq(10090)"', result = 'assembly')
dplyr::bind_rows(lapply(res, "[[", "chromosomes"), .id = "id")

# coding_release
res <- en_search(query = '"tax_tree(7147) AND dataclass="STD""', result = 'coding_release', limit = 5)
vapply(res, "[[", "", "desc")

# coding_update
res <- en_search(query = '"tax_tree(7147) AND dataclass="STD""', result = 'coding_update', limit = 5)
res$AAA08526$refs
res$AAA08526$features

## failing
# en_search(query = '"tax_tree(7147) AND dataclass="STD"')
}
}
\references{
\url{http://www.ebi.ac.uk/ena/browse/search-rest}
}

